remove the focus
This is part1 of a fix for https://github.com/joomla/joomla-cms/issues/25348

We should not be changing the focus to the alert element. If we do then that forces the user to close the alert and they have lost their place in the document. 

Part2 which will be in the core layout is to add the attribute aria-live to the message container. This will ensure that the message will be announced by the screenreader even though we don't give it focus